requirements/base.txt
click==6.7
Flask-Migrate==2.2.1
Flask-RQ2==18.3
Flask-SQLAlchemy==2.5.1
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==2.0.1
pydantic==1.8.2
python-baseconv==1.2.2
python-dateutil==2.7.3
python-dotenv==0.9.1
requests==2.19.1
rq-dashboard==0.6.1
simplejson==3.16.0
Werkzeug==0.14.1
